Cleanup the whole proto section
